Refactor component items
